requirements.txt
streamlit
streamlit-authenticator
PyYAML